refactor: demote CompressedCodec to bench-only utility per #2831
Team direction is option 5 (connections output CompressedImage directly);
the transport wrapper was rejected as public API. Keep it inside
tool_replay_bench.py so raw-vs-codec benchmark cells stay reproducible,
move its tests alongside, drop the unitree-go2-compressed-image blueprint.
